sql/parser: move pg builtins to their own file
